# openapi.yaml
openapi: 3.0.3
info:
  title: AI数字名片 API
  description: AI数字名片后端服务 - 模块化架构
  version: 2.0.0
  contact:
    name: AI数字名片团队
servers:
- url: http://localhost:8201
  description: 本地开发
- url: https://api.example.com
  description: 生产环境
paths:
...
  /api/brochures/template/{purpose}:
    get:
      summary: 获取用途推荐模板
      tags:
      - 画册
      parameters:
      - name: purpose
        in: path
        required: true
        schema:
          type: string
          enum:
          - partner
          - client
          - investor
          - supplier
      responses:
        '200':
          description: Successful Response
  /api/brochures:
    post:
      summary: 创建画册
      tags:
      - 画册
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrochureResponse'
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrochureCreate'
    get:
      summary: 获取画册列表
      tags:
      - 画册
      parameters:
      - name: user_id
        in: query
        required: false
        schema:
          type: integer
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          default: 0
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
      responses:
        '200':
          description: 画册列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BrochureResponse'
  /api/brochures/{brochure_id}:
    get:
      summary: 获取画册详情
      tags:
      - 画册
      parameters:
      - name: brochure_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrochureResponse'
    put:
      summary: 更新画册
      tags:
      - 画册
      parameters:
      - name: brochure_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrochureResponse'
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrochureUpdate'
    delete:
      summary: 删除画册
      tags:
      - 画册
      parameters:
      - name: brochure_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 删除成功
        '204':
          description: 无内容
      security:
      - BearerAuth: []
  /api/brochures/share/{share_token}:
    get:
      summary: 通过分享token获取画册(公开)
      tags:
      - 画册
      parameters:
      - name: share_token
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrochureResponse'
  /api/brochures/{brochure_id}/publish:
    post:
      summary: 发布画册
      tags:
      - 画册
      parameters:
      - name: brochure_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrochureResponse'
      security:
      - BearerAuth: []
  /api/brochures/{brochure_id}/refresh-token:
    post:
      summary: 刷新分享token
      tags:
      - 画册
      parameters:
      - name: brochure_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Successful Response
      security:
      - BearerAuth: []
  /api/brochures/smart-search:
    post:
      summary: 智能搜索名片
      tags:
      - 画册
      parameters: []
      responses:
        '200':
          description: Successful Response
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmartSearchQuery'
  /api/brochures/upload-video:
    post:
      summary: 上传视频文件
      tags:
      - 画册
      parameters: []
      responses:
        '200':
          description: Successful Response
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
  /api/brochures/batch-import:
    post:
      summary: 批量导入名片(CSV)
      tags:
      - 画册
      parameters: []
      responses:
        '200':
          description: Successful Response
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
  /api/brochures/batch-export:
    post:
      summary: 批量导出名片为CSV
      tags:
      - 画册
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Successful Response
      security:
      - BearerAuth: []
...
  /api/brochure/smart-search:
    post:
      summary: 智能搜索(别名路由)
      tags:
      - 画册别名
      parameters: []
      responses:
        '200':
          description: Successful Response
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmartSearchQuery'
  /api/brochure/template/{purpose}:
    get:
      summary: 获取模板(别名路由)
      tags:
      - 画册别名
      parameters:
      - name: purpose
        in: path
        required: true
        schema:
          type: string
          enum:
          - partner
          - client
          - investor
          - supplier
      responses:
        '200':
          description: Successful Response
...
  /api/v1/admin/brochures:
    get:
      summary: 管理员-名片列表
      tags:
      - Admin Panel
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
      - name: per_page
        in: query
        required: false
        schema:
          type: integer
          default: 20
      responses:
        '200':
          description: Successful Response
      security:
      - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT令牌, 通过 /api/auth/login 获取
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key, 通过 /api/api-keys 创建
  schemas:
...
    BrochureCreate:
      type: object
      required:
      - title
      - pages
      properties:
        title:
          type: string
        cover:
          type: string
        purpose:
          type: string
          enum:
          - partner
          - client
          - investor
          - supplier
        album_meta:
          type: object
        pages:
          type: array
          items:
            $ref: '#/components/schemas/PageSchema'
    BrochureUpdate:
      type: object
      properties:
        title:
          type: string
        cover:
          type: string
        purpose:
          type: string
        album_meta:
          type: object
        pages:
          type: array
          items:
            $ref: '#/components/schemas/PageSchema'
    BrochureResponse:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        title:
          type: string
        cover:
          type: string
        purpose:
          type: string
        status:
          type: string
        share_token:
          type: string
        view_count:
          type: integer
        pages_count:
          type: integer
        album_meta:
          type: object
        pages:
          type: array
          items:
            $ref: '#/components/schemas/PageSchema'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PageSchema:
      type: object
      properties:
        id:
          type: integer
        sort_order:
          type: integer
        content_type:
          type: string
        content:
          type: string
        image_url:
          type: string
        media_url:
          type: string
        ai_summary:
          type: string
    SmartSearchQuery:
      type: object
      required:
      - query
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 500
        top_k:
          type: integer
          default: 10
        mode:
          type: string
          enum:
          - hybrid
          - vector
          - keyword
          default: hybrid
        min_score:
          type: number
          default: 0.0
...